Backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
PyMySQL==1.0.3
Flask-Migrate==4.0.4
Flask-Caching==2.0.2
gunicorn==21.2.0
flask_wtf
python-dotenv
cryptography
pyjwt==2.9.0
flask_cors==5.0.0


redis==5.0.0
rq
